requirements.txt
gspread==6.1.4
langchain_community==0.3.8
langchain_core==0.3.21
langchain_experimental==0.3.3
langchain_google_genai==2.0.6
langchain_groq==0.2.1
langchain_nomic==0.1.4
langchain_ollama==0.2.0
langchain-qdrant==0.2.0
neo4j==5.25.0
oauth2client==4.1.3
pydantic==2.10.2
python-dotenv==1.0.1
streamlit==1.40.1
llama-index-core==0.11.23
llama-parse==0.5.14
llama-index-readers-file==0.3.0
RapidFuzz==3.10.1
